fixed the spaces with edit discussion
